+++ app/src/main/java/a75f/io/renatus/OTAUpdateHandlerService.java
@@ -96,7 +96,7 @@ public class OTAUpdateHandlerService extends Service {
                 @Override
                 public void onTick(long millisUntilFinished) {
                     Log.d(TAG, "[TIMER] Update will time out in " +
-                            Math.ceil(millisUntilFinished / 1000) + "s");
+                            Math.ceil((double) millisUntilFinished / 1000) + "s");
                 }
 
                 @Override
